First bash at updating classimp to use new cffi type style
The with-foreign-slots macro is now a pain as it gets the slot value, which for
structs is the plist. This sucks when we need a pointer so I have added the utils
file which has the with-foreign-slots* macro. This allows you to specify a name
to get the value or (:pointer name) to get the pointer. I'm updating wrappers
to use this exclusively, this seems to be the easiest way to get the desired
functionality.

Ok so I can load files again now but still have massive numbers of struct warnings to do with vector3d...weird.

All cstructs should now have new type declarations... Now I need to check defcfuns

haha lots of the cstructs were enums so had to undo some clumsy work (what a newbie!)
Anyhoo I have updated the defcfuns and the occasions in wrappers.lisp where bare names
were used.... its loads files now without drama!... they may not loaded correctly but that
will be revealed.

Create README.md

* Found more cffi related issues (missing :struct)
* Made sure all (format t ..)  expressions obey the
  *translate-verbose* flag

Bare structs resolved. Readme updated